0.1.0 -- JamFix release
For the 10-day Spring Lisp Game Jam, Winny and Aliasing made a game
called Super Rogue, which got 4th place in the competition. This is a
fusion of SuperHot and the classic Rogue, which the RogueLike genre is
named after. This release is a gameplay proof-of-concept, with only 3
enemies -- Grid Bug, Killer Tomato, and Leprechaun, and very basic
mechanics. There is no inventory system, and there are no items other
than money. However, most of the bedrock for a bigger, better game has
been laid, and a fun demo now exists for you to play. This is the
official post-jam release of the Jam game, which fixes the bug of
bullets going thru walls, and makes the game's icon and banner a bit
more appealing, as well as adding a version number to the title screen.